refactor(og-image): move og image seo functions into its own composable
